requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.1.8
distro==1.9.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.77.0
pip==25.1.1
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
setuptools==65.5.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing_extensions==4.13.2
typing-inspection==0.4.0
uvicorn==0.34.2
